feat: build files
