+++ packages/worker-v4/src/index.ts
@@ -1,62 +1,172 @@
 import type { WorkerRunnerOptions } from 'vitest-vscode-shared'
-import v8 from 'node:v8'
-import { createWorkerRPC, WorkerWSEventEmitter } from 'vitest-vscode-shared'
-import { WebSocket } from 'ws'
-import { initVitest } from './init'
+import type { WorkerWSEventEmitter } from 'vitest-vscode-shared'
+import type { TestUserConfig } from 'vitest/dist/node.js'
+import { Console } from 'node:console'
+import { randomUUID } from 'node:crypto'
+import { tmpdir } from 'node:os'
+import { join } from 'node:path'
+import { Writable } from 'node:stream'
+import { VSCodeReporter } from './reporter'
 import { ExtensionWorker } from './worker'
 
-// this is the file that will be executed with "node <path>"
+export async function initVitest(
+  vitestModule: typeof import('vitest/node'),
+  data: WorkerRunnerOptions,
+  emitter: WorkerWSEventEmitter,
+) {
+  const meta = data.meta
+  const reporter = new VSCodeReporter({
+    setupFilePath: meta.setupFilePath,
+  })
 
-const emitter = new WorkerWSEventEmitter(
-  new WebSocket(process.env.VITEST_WS_ADDRESS!),
-)
+  let stdout: Writable | undefined
+  let stderr: Writable | undefined
 
-emitter.on('message', async function onMessage(message: any) {
-  if (emitter.name === 'ws') {
-    message = JSON.parse(message.toString())
+  if (meta.shellType === 'terminal' && !meta.hasShellIntegration) {
+    stdout = new Writable({
+      write(chunk, __, callback) {
+        const log = chunk.toString()
+        if (reporter.rpc) {
+          reporter.rpc.onProcessLog('stdout', log).catch(() => {})
+        }
+        process.stdout.write(log)
+        callback()
+      },
+    })
+
+    stderr = new Writable({
+      write(chunk, __, callback) {
+        const log = chunk.toString()
+        if (reporter.rpc) {
+          reporter.rpc.onProcessLog('stderr', log).catch(() => {})
+        }
+        process.stderr.write(log)
+        callback()
+      },
+    })
+    globalThis.console = new Console(stdout, stderr)
   }
 
-  if (message.type === 'init') {
-    emitter.off('message', onMessage)
-    const data = message as WorkerRunnerOptions
-
-    try {
-      const { reporter, vitest, configs, workspaceSource } = await initVitest(
-        data.meta,
-        data.debug
-          ? {
-              disableConsoleIntercept: true,
-              fileParallelism: false,
-              testTimeout: 0,
-              hookTimeout: 0,
-            }
-          : {},
-      )
-
-      const rpc = createWorkerRPC(
-        new ExtensionWorker(
-          vitest,
-          data.debug,
-          data.astCollect,
-          emitter,
-          data.meta.finalCoverageFileName,
-        ),
+  const pnpExecArgv = meta.pnpApi && meta.pnpLoader
+    ? [
+        '--require',
+        meta.pnpApi,
+        '--experimental-loader',
+        meta.pnpLoader,
+      ]
+    : undefined
+  const args = meta.arguments
+    ? vitestModule.parseCLI(meta.arguments, {
+      allowUnknownOptions: false,
+    }).options
+    : {}
+  const options = data.debug
+    ? {
+        disableConsoleIntercept: true,
+        fileParallelism: false,
+        testTimeout: 0,
+        hookTimeout: 0,
+      }
+    : {}
+  const cliOptions: TestUserConfig = {
+    config: meta.configFile,
+    ...(meta.workspaceFile ? { workspace: meta.workspaceFile } : {}),
+    ...args,
+    ...options,
+    watch: true,
+    api: false,
+    // @ts-expect-error private property
+    reporter: undefined,
+    reporters: [reporter],
+    ui: false,
+    includeTaskLocation: true,
+    poolOptions: meta.pnpApi && meta.pnpLoader
+      ? {
+          threads: {
+            execArgv: pnpExecArgv,
+          },
+          forks: {
+            execArgv: pnpExecArgv,
+          },
+          vmForks: {
+            execArgv: pnpExecArgv,
+          },
+          vmThreads: {
+            execArgv: pnpExecArgv,
+          },
+        }
+      : {},
+  }
+  const vitest = await vitestModule.createVitest(
+    'test',
+    cliOptions,
+    {
+      server: {
+        middlewareMode: true,
+        watch: null,
+      },
+      plugins: [
         {
-          on(listener) {
-            emitter.on('message', listener)
+          name: 'vitest:vscode-extension',
+          config(userConfig) {
+            const testConfig = userConfig.test ?? {}
+            const coverageOptions = testConfig.coverage ?? {}
+            if (coverageOptions.provider !== 'custom') {
+              const reporters = Array.isArray(coverageOptions.reporter) ? coverageOptions.reporter : [coverageOptions.reporter]
+              const jsonReporter = reporters.find(r => r && r[0] === 'json')
+              const jsonReporterOptions = typeof jsonReporter?.[1] === 'object' ? jsonReporter[1] : {}
+              coverageOptions.reporter = [
+                ['json', { ...jsonReporterOptions, file: meta.finalCoverageFileName }],
+              ]
+            }
+            return {
+              test: {
+                coverage: {
+                  reportOnFailure: true,
+                  reportsDirectory: join(tmpdir(), `vitest-coverage-${randomUUID()}`),
+                },
+              },
+            }
           },
-          post(message) {
-            emitter.send(message)
+          configResolved(config) {
+            // stub a server so Vite doesn't start a websocket connection,
+            // because we don't need it in the extension and it messes up Vite dev command
+            config.server.hmr = {
+              server: {
+                on: () => {},
+                off: () => {},
+              } as any,
+            }
           },
-          serialize: v8.serialize,
-          deserialize: v => v8.deserialize(Buffer.from(v) as any),
         },
+      ],
+    },
+    {
+      stderr,
+      stdout,
+    },
+  )
+  await (vitest as any).report('onInit', vitest)
+  const configs = [
+    vitest.getRootProject(),
+    ...vitest.projects,
+  ].map(p => p.vite.config.configFile).filter(c => c != null)
+  const workspaceSource: string | false = (vitest.config.projects != null)
+    ? vitest.vite.config.configFile || false
+    : false
+  return {
+    vitest,
+    reporter,
+    workspaceSource,
+    configs: Array.from(new Set(configs)),
+    meta,
+    createWorker() {
+      return new ExtensionWorker(
+        vitest,
+        data.debug,
+        data.astCollect,
+        emitter,
       )
-      reporter.initRpc(rpc)
-      emitter.ready(configs, workspaceSource)
-    }
-    catch (err: any) {
-      emitter.error(err)
-    }
+    },
   }
-})
+}
